Move sidecar instance
