+++ analytics-sidecar-client/src/main/java/org/apache/cassandra/sidecar/client/HttpResponse.java
@@ -22,8 +22,6 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 import java.util.Map;
 
-import org.apache.cassandra.sidecar.common.client.SidecarInstance;
-
 /**
  * Represents the HTTP response received from the remote Sidecar service
  */
