+++ analytics-sidecar-client/src/main/java/org/apache/cassandra/sidecar/client/HttpResponseImpl.java
@@ -23,8 +23,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import org.apache.cassandra.sidecar.common.client.SidecarInstance;
-
 /**
  * A simple implementation of the {@link HttpResponse}
  */
