+++ analytics-sidecar-client/src/main/java/org/apache/cassandra/sidecar/client/SidecarClient.java
@@ -36,7 +36,6 @@ import org.apache.cassandra.sidecar.client.retry.RetryPolicy;
 import org.apache.cassandra.sidecar.client.retry.RunnableOnStatusCodeRetryPolicy;
 import org.apache.cassandra.sidecar.client.selection.InstanceSelectionPolicy;
 import org.apache.cassandra.sidecar.client.selection.RandomInstanceSelectionPolicy;
-import org.apache.cassandra.sidecar.common.client.SidecarInstance;
 import org.apache.cassandra.sidecar.common.request.AbortRestoreJobRequest;
 import org.apache.cassandra.sidecar.common.request.AllServicesConfigRequest;
 import org.apache.cassandra.sidecar.common.request.CreateRestoreJobRequest;
