+++ analytics-sidecar-client/src/main/java/org/apache/cassandra/sidecar/client/selection/OrderedInstanceSelectionPolicy.java
@@ -21,7 +21,7 @@ package org.apache.cassandra.sidecar.client.selection;
 import java.util.Iterator;
 
 import org.apache.cassandra.sidecar.client.SidecarInstancesProvider;
-import org.apache.cassandra.sidecar.common.client.SidecarInstance;
+import org.apache.cassandra.sidecar.client.SidecarInstance;
 import org.jetbrains.annotations.NotNull;
 
 /**
