+++ analytics-sidecar-client/src/test/java/org.apache.cassandra.sidecar.client/SidecarInstanceImplTest.java
@@ -18,9 +18,6 @@
 
 package org.apache.cassandra.sidecar.client;
 
-import org.apache.cassandra.sidecar.common.client.SidecarInstance;
-import org.apache.cassandra.sidecar.common.client.SidecarInstanceImpl;
-
 /**
  * Unit tests for the {@link SidecarInstanceImpl} class
  */
