+++ analytics-sidecar-client/src/test/java/org.apache.cassandra.sidecar.client/selection/OrderedInstanceSelectionPolicyTest.java
@@ -28,7 +28,7 @@ import org.junit.jupiter.api.Test;
 
 import org.apache.cassandra.sidecar.client.SidecarInstancesProvider;
 import org.apache.cassandra.sidecar.client.SimpleSidecarInstancesProvider;
-import org.apache.cassandra.sidecar.common.client.SidecarInstance;
+import org.apache.cassandra.sidecar.client.SidecarInstance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
