+++ analytics-sidecar-vertx-client/src/test/java/org/apache/cassandra/sidecar/client/VertxSidecarClientTest.java
@@ -24,7 +24,6 @@ import java.util.List;
 import io.vertx.core.Vertx;
 import org.apache.cassandra.sidecar.client.retry.ExponentialBackoffRetryPolicy;
 import org.apache.cassandra.sidecar.client.retry.RetryPolicy;
-import org.apache.cassandra.sidecar.common.client.SidecarInstanceImpl;
 
 /**
  * Unit tests for the {@link SidecarClient} using vertx
